Improve comps group support
New file dnfdragora/compsicons.py, similar to groupicons.py, but for
comps.

New tool tools/gen-comps-category-list.sh: Generates a map from category
name to ID for dnfdragora/compsicons.py, as a workaround for
timlau/dnf-daemon#9.

dnfdragora/ui.py (mainGui._getAllGroupIDList): Remember the UI names for
                                               each ID in an
                                               id_to_name_map.
                 (mainGui._fillGroupTree): Handle comps separately, use
                                           compsicons.py to get the
                                           icons (#2) and use the
                                           id_to_name_map to get the UI
                                           name (#3).

Fixes #2.
Fixes #3.

The only remaining issue is that the UI names are not translated, which
I filed as timlau/dnf-daemon#10 because Yumex-DNF is also affected (and
thus it is clearly dnfdaemon's fault).